v0.0.1 activity.exit & activity.exit.exitArgumentsType ⬇︎
The object activity.exit is a function to exit the process giving a message and a stack trace.
it prints to stderr.

The object activity.exit.exitArgumentType is a function to type checking the arguments passed to activity.exit
it prints to stderr and exits with errors.type_checking code = 3 should be considered `private` property of activity.exit